FITRUN-118 refactor: ApiResponse 패키지 변경
