+++ src/test/kotlin/com/yapp/yapp/document/auth/AuthDocumentTest.kt
@@ -4,7 +4,8 @@ import com.yapp.yapp.auth.api.request.LoginRequest
 import com.yapp.yapp.auth.api.response.LoginResponse
 import com.yapp.yapp.auth.api.response.TokenResponse
 import com.yapp.yapp.auth.infrastructure.provider.apple.AppleFeignClient
-import com.yapp.yapp.common.ApiResponse
+import com.yapp.yapp.auth.infrastructure.provider.kakao.KakaoFeignClient
+import com.yapp.yapp.common.web.ApiResponse
 import com.yapp.yapp.document.Tag
 import com.yapp.yapp.document.support.BaseDocumentTest
 import com.yapp.yapp.support.fixture.IdTokenFixture
